# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JARVIS AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiVoiceMessageInput {
  /** Base64-encoded audio data */
  audio: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface Memory {
  id: number;
  content: string;
  category: string;
  createdAt: string;
}

export interface MemoryInput {
  content: string;
  category: string;
}

export type NoteType = typeof NoteType[keyof typeof NoteType];


export const NoteType = {
  note: 'note',
  reminder: 'reminder',
  automation: 'automation',
} as const;

export interface Note {
  id: number;
  title: string;
  content: string;
  type: NoteType;
  completed: boolean;
  /** @nullable */
  reminderAt?: string | null;
  createdAt: string;
}

export type NoteInputType = typeof NoteInputType[keyof typeof NoteInputType];


export const NoteInputType = {
  note: 'note',
  reminder: 'reminder',
  automation: 'automation',
} as const;

export interface NoteInput {
  title: string;
  content: string;
  type: NoteInputType;
  /** @nullable */
  reminderAt?: string | null;
}

export interface NoteUpdate {
  title?: string;
  content?: string;
  completed?: boolean;
  /** @nullable */
  reminderAt?: string | null;
}

export type UserSettingsPersonality = typeof UserSettingsPersonality[keyof typeof UserSettingsPersonality];


export const UserSettingsPersonality = {
  jarvis: 'jarvis',
  friday: 'friday',
  karen: 'karen',
  custom: 'custom',
} as const;

export type UserSettingsVoiceGender = typeof UserSettingsVoiceGender[keyof typeof UserSettingsVoiceGender];


export const UserSettingsVoiceGender = {
  alloy: 'alloy',
  echo: 'echo',
  fable: 'fable',
  onyx: 'onyx',
  nova: 'nova',
  shimmer: 'shimmer',
} as const;

export type UserSettingsTheme = typeof UserSettingsTheme[keyof typeof UserSettingsTheme];


export const UserSettingsTheme = {
  blue: 'blue',
  red: 'red',
  green: 'green',
  gold: 'gold',
} as const;

export interface UserSettings {
  id: number;
  assistantName: string;
  personality: UserSettingsPersonality;
  /** @nullable */
  customPersonality?: string | null;
  voiceEnabled: boolean;
  voiceGender: UserSettingsVoiceGender;
  theme: UserSettingsTheme;
  createdAt: string;
}

export type UserSettingsUpdatePersonality = typeof UserSettingsUpdatePersonality[keyof typeof UserSettingsUpdatePersonality];


export const UserSettingsUpdatePersonality = {
  jarvis: 'jarvis',
  friday: 'friday',
  karen: 'karen',
  custom: 'custom',
} as const;

export type UserSettingsUpdateVoiceGender = typeof UserSettingsUpdateVoiceGender[keyof typeof UserSettingsUpdateVoiceGender];


export const UserSettingsUpdateVoiceGender = {
  alloy: 'alloy',
  echo: 'echo',
  fable: 'fable',
  onyx: 'onyx',
  nova: 'nova',
  shimmer: 'shimmer',
} as const;

export type UserSettingsUpdateTheme = typeof UserSettingsUpdateTheme[keyof typeof UserSettingsUpdateTheme];


export const UserSettingsUpdateTheme = {
  blue: 'blue',
  red: 'red',
  green: 'green',
  gold: 'gold',
} as const;

export interface UserSettingsUpdate {
  assistantName?: string;
  personality?: UserSettingsUpdatePersonality;
  /** @nullable */
  customPersonality?: string | null;
  voiceEnabled?: boolean;
  voiceGender?: UserSettingsUpdateVoiceGender;
  theme?: UserSettingsUpdateTheme;
}

export interface DashboardStats {
  totalConversations: number;
  totalMessages: number;
  totalMemories: number;
  totalNotes: number;
  activeReminders: number;
}