# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JARVIS AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { NoteType } from './noteType';

export interface Note {
  id: number;
  title: string;
  content: string;
  type: NoteType;
  completed: boolean;
  /** @nullable */
  reminderAt?: Date | null;
  createdAt: Date;
}